+++ mapper/lib/src/model/enum.dart
@@ -3,23 +3,48 @@
 /// enum ABC {A, B, C}, etc.
 abstract class Enum {}
 
-/// Enum descriptor, values, mappings, etc
+/// Enum descriptor, defines Enum possible values, mappings, defaultValue, case sensitivity, etc
 abstract class IEnumDescriptor {
-  Iterable? values;
+  Iterable values = [];
   dynamic defaultValue;
-  Map? mapping;
+  bool? caseInsensitive;
+  Map mapping = {};
 }
 
-/// EnumDescriptor
+/// Enum descriptor, defines Enum possible values, mappings, defaultValue, case sensitivity, etc
 class EnumDescriptor implements IEnumDescriptor {
+  /// Defines a mapping for enum values, key is the enum value, value is the target mapping value
+  /// Example:
+  /// EnumDescriptor(
+  ///           values: RecordType.values,
+  ///           mapping: <RecordType, String>{
+  ///             RecordType.asset: 'Asset',
+  ///             RecordType.series: 'Series'
+  ///           })
   @override
-  Map? mapping;
+  Map mapping;
 
+  /// Defines possible enum values
+  /// Example:
+  /// EnumDescriptor(values: RecordType.values)
   @override
-  Iterable? values;
+  Iterable values;
 
+  /// Defines possible enum values
+  /// Example:
+  /// EnumDescriptor(defaultValue: RecordType.asset)
   @override
   dynamic defaultValue;
 
-  EnumDescriptor({this.values, this.mapping, this.defaultValue});
+  /// Defines case sensitivity for string based enum values
+  /// Example:
+  /// EnumDescriptor(values: RecordType.values, caseInsensitive: true)
+  @override
+  bool? caseInsensitive;
+
+  EnumDescriptor(
+      {this.values = const [],
+      this.mapping = const {},
+      this.defaultValue,
+      this.caseInsensitive});
 }